requirements.txt
agno
python-dotenv
pydantic
requests
openai
sqlalchemy
redis
fastapi
uvicorn[standard]
httpx
